requirements.txt
selenium
pytest
pytest-rerunfailures
pytest-xdist
allure-pytest
PyYAML
requests
allure-python-commons